LayoutTests/platform/glib/imported/w3c/web-platform-tests/fetch/private-network-access/preflight-cache.https.window-expected.txt


FAIL private to local: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Load failed"
FAIL public to local: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Load failed"
FAIL public to private: success. assert_equals: error expected (undefined) undefined but got (string) "TypeError: Load failed"
